ajustes na query de encontrar conta, construtores do jpa, rota d depositar padrao em usuario, mudanca de tipo

+++ picpay_simplificado/src/main/java/com/desafio/picpay_simplificado/picpay_simplificado/Interface/EnvioDeTransferencias.java
@@ -2,6 +2,6 @@ package com.desafio.picpay_simplificado.picpay_simplificado.Interface;
 
 public interface EnvioDeTransferencias {
 
-    public void transfere(float valorDaTransferencia) throws Exception;
+    public void transfere(double valorDaTransferencia) throws Exception;
 
 }
