+++ picpay_simplificado/src/main/java/com/desafio/picpay_simplificado/picpay_simplificado/Interface/RecebeDepositos.java
@@ -2,6 +2,6 @@ package com.desafio.picpay_simplificado.picpay_simplificado.Interface;
 
 public interface RecebeDepositos {
 
-    public void deposita(float valorADepositar) throws Exception;
+    public void deposita(Double valorADepositar) throws Exception;
 
 }
